Swap Active Job log subscriber safely
